+++ docker/consumer/consumer.py
@@ -2,7 +2,6 @@ from kafka import KafkaConsumer
 import json
 import psycopg2
 import os
-from datetime import datetime
 from dotenv import load_dotenv
 import logging 
 from sys import stdout
