+++ docker/producer/requirements.txt
@@ -18,3 +18,4 @@ kafka-python==2.0.2
 python-dotenv==0.21.0
 urllib3==1.26.7
 requests==2.26.0 
+selenium==3.141.0
